Final working code before correcting the forward coefficients
The valuse of CTYPEn is still hardcoded but it's the only
way it works. Included loops in gal_wcs_add_sipkeywords, though the size
of characters of some of the cards keywords is increased like for axis[9]
pcaxis[10] which should only be 8 char wide but shouldn't be a problem
cause 8 bit integer is used so they shouldn't overflow beyond the 8
character limit.