Minor modifications in some functions and initial during-dev test
Some minor modifications were made to some functions removing
redundant calls to reading input WCS and data. Also an inital
implementation of the high-level function was added in the end.

Finally, a new 'for-during-dev.c' file is added which you can in a
call to 'astbuildprog' in Gnuastro's 'tests/during-dev.sh'.